+++ framework/db/Connection.php
@@ -281,9 +281,9 @@ class Connection extends Component
         'mysql' => 'yii\db\mysql\Schema', // MySQL
         'sqlite' => 'yii\db\sqlite\Schema', // sqlite 3
         'sqlite2' => 'yii\db\sqlite\Schema', // sqlite 2
-        'sqlsrv' => 'yii\db\mssql\Schema', // newer MSSQL driver on MS Windows hosts
+        'sqlsrv' => 'yii\db\mssql\Schema', // newer MSSQL driver
         'oci' => 'yii\db\oci\Schema', // Oracle driver
-        'mssql' => 'yii\db\mssql\Schema', // older MSSQL driver on MS Windows hosts
+        'mssql' => 'yii\db\mssql\Schema', // older MSSQL driver
         'dblib' => 'yii\db\mssql\Schema', // dblib drivers on GNU/Linux (and maybe other OSes) hosts
     ];
     /**
@@ -308,10 +308,10 @@ class Connection extends Component
         'mysql' => 'yii\db\Command', // MySQL
         'sqlite' => 'yii\db\sqlite\Command', // sqlite 3
         'sqlite2' => 'yii\db\sqlite\Command', // sqlite 2
-        'sqlsrv' => 'yii\db\Command', // newer MSSQL driver on MS Windows hosts
+        'sqlsrv' => 'yii\db\mssql\Command', // newer MSSQL driver on MS Windows hosts
         'oci' => 'yii\db\oci\Command', // Oracle driver
-        'mssql' => 'yii\db\Command', // older MSSQL driver on MS Windows hosts
-        'dblib' => 'yii\db\Command', // dblib drivers on GNU/Linux (and maybe other OSes) hosts
+        'mssql' => 'yii\db\mssql\Command', // older MSSQL driver on MS Windows hosts
+        'dblib' => 'yii\db\mssql\Command', // dblib drivers on GNU/Linux (and maybe other OSes) hosts
     ];
     /**
      * @var bool whether to enable [savepoint](https://en.wikipedia.org/wiki/Savepoint).